refactor: activities into ui folder
