+++ app/src/main/java/com/devfiveurjc/agendaly/ui/ActionTasksActivity.java
@@ -1,9 +1,11 @@
-package com.devfiveurjc.agendaly;
+package com.devfiveurjc.agendaly.ui;
 
 import android.os.Bundle;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.devfiveurjc.agendaly.R;
 
 import java.util.ArrayList;
 import java.util.Calendar;
