+++ app/src/main/java/com/devfiveurjc/agendaly/ui/CalendarActivity.java
@@ -1,10 +1,12 @@
-package com.devfiveurjc.agendaly;
+package com.devfiveurjc.agendaly.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.devfiveurjc.agendaly.R;
 
 public class CalendarActivity extends AppCompatActivity {
     @Override
