+++ app/src/main/java/com/devfiveurjc/agendaly/ui/Task.java
@@ -1,4 +1,4 @@
-package com.devfiveurjc.agendaly;
+package com.devfiveurjc.agendaly.ui;
 
 import java.util.Calendar;
