+++ app/src/main/java/com/devfiveurjc/agendaly/ui/ViewTasksActivity.java
@@ -1,4 +1,4 @@
-package com.devfiveurjc.agendaly;
+package com.devfiveurjc.agendaly.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.devfiveurjc.agendaly.R;
 
 import java.util.ArrayList;
 import java.util.Calendar;
