+++ app/src/main/res/layout/activity_action_tasks.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/personalizado"
-    tools:context=".ActionTasksActivity">
+    tools:context=".ui.ActionTasksActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar5"
