+++ app/src/main/res/layout/activity_calendar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/personalizado"
-    tools:context=".ViewTasksActivity">
+    tools:context=".ui.ViewTasksActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar3"
