+++ app/src/main/res/layout/activity_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/personalizado"
-    tools:context=".MenuActivity">
+    tools:context=".ui.MenuActivity">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/exit"
