+++ app/src/main/res/layout/activity_new_task.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/personalizado"
-    tools:context=".NewTaskActivity">
+    tools:context=".ui.NewTaskActivity">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/button2"
