+++ app/src/main/res/layout/activity_view_tasks.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/personalizado"
-    tools:context=".ViewTasksActivity">
+    tools:context=".ui.ViewTasksActivity">
 
     <ScrollView
         android:id="@+id/scrollView2"
